# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an API description for [Moira Alert API](https://moira.readthedocs.io/en/latest/overview.html).",
        "title": "Moira Alert",
        "contact": {
            "name": "Contact Moira Team",
            "email": "opensource@skbkontur.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "master"
    },
    "basePath": "/api",
    "paths": {
...
        "/trigger": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trigger"
                ],
                "summary": "Get all triggers",
                "operationId": "get-all-triggers",
                "responses": {
                    "200": {
                        "description": "Fetched all triggers",
                        "schema": {
                            "$ref": "#/definitions/dto.TriggersList"
                        }
                    },
                    "422": {
                        "description": "Render error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorRenderExample"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorInternalServerExample"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trigger"
                ],
                "summary": "Create a new trigger",
                "operationId": "create-trigger",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User session token",
                        "name": "x-webauth-user",
                        "in": "header"
                    },
                    {
                        "description": "Trigger data",
                        "name": "trigger",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.Trigger"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trigger created successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.SaveTriggerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request from client",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorInvalidRequestExample"
                        }
                    },
                    "422": {
                        "description": "Render error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorRenderExample"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorInternalServerExample"
                        }
                    },
                    "503": {
                        "description": "Remote server unavailable",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorRemoteServerUnavailableExample"
                        }
                    }
                }
            }
        },
        "/trigger/check": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trigger"
                ],
                "summary": "Validates trigger target",
                "operationId": "trigger-check",
                "parameters": [
                    {
                        "description": "Trigger data",
                        "name": "trigger",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.Trigger"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Validation is done, see response body for validation result",
                        "schema": {
                            "$ref": "#/definitions/dto.TriggerCheckResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request from client",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorInvalidRequestExample"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorInternalServerExample"
                        }
                    }
                }
            }
        },
...
        "/trigger/search": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trigger"
                ],
                "summary": "Search triggers. Replaces the deprecated `page` path",
                "operationId": "search-triggers",
                "parameters": [
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Only include problems",
                        "name": "onlyProblems",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "cpu",
                        "description": "Search text",
                        "name": "text",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Page number",
                        "name": "p",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Page size",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Create pager",
                        "name": "createPager",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "bcba82f5-48cf-44c0-b7d6-e1d32c64a88c",
                        "description": "Pager ID",
                        "name": "pagerID",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully fetched matching triggers",
                        "schema": {
                            "$ref": "#/definitions/dto.TriggersList"
                        }
                    },
                    "400": {
                        "description": "Bad request from client",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorInvalidRequestExample"
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorNotFoundExample"
                        }
                    },
                    "422": {
                        "description": "Render error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorRenderExample"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorInternalServerExample"
                        }
                    }
                }
            }
        },
        "/trigger/{triggerID}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trigger"
                ],
                "summary": "Get an existing trigger",
                "operationId": "get-trigger",
                "parameters": [
                    {
                        "type": "string",
                        "default": "bcba82f5-48cf-44c0-b7d6-e1d32c64a88c",
                        "description": "Trigger ID",
                        "name": "triggerID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trigger data",
                        "schema": {
                            "$ref": "#/definitions/dto.Trigger"
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorNotFoundExample"
                        }
                    },
                    "422": {
                        "description": "Render error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorRenderExample"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorInternalServerExample"
                        }
                    }
                }
            },
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trigger"
                ],
                "summary": "Update existing trigger",
                "operationId": "update-trigger",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User session token",
                        "name": "x-webauth-user",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "default": "bcba82f5-48cf-44c0-b7d6-e1d32c64a88c",
                        "description": "Trigger ID",
                        "name": "triggerID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Trigger data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.Trigger"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated trigger",
                        "schema": {
                            "$ref": "#/definitions/dto.SaveTriggerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request from client",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorInvalidRequestExample"
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorNotFoundExample"
                        }
                    },
                    "422": {
                        "description": "Render error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorRenderExample"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorInternalServerExample"
                        }
                    },
                    "503": {
                        "description": "Remote server unavailable",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorRemoteServerUnavailableExample"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "trigger"
                ],
                "summary": "Remove trigger",
                "operationId": "remove-trigger",
                "parameters": [
                    {
                        "type": "string",
                        "default": "bcba82f5-48cf-44c0-b7d6-e1d32c64a88c",
                        "description": "Trigger ID",
                        "name": "triggerID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Resource not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorNotFoundExample"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorInternalServerExample"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.ErrorInternalServerExample": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "server error during request handling"
                },
                "status": {
                    "type": "string",
                    "example": "Internal Server Error"
                }
            }
        },
        "api.ErrorInvalidRequestExample": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "resource with the ID does not exist"
                },
                "status": {
                    "type": "string",
                    "example": "Invalid request"
                }
            }
        },
        "api.ErrorNotFoundExample": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "resource with ID '66741a8c-c2ba-4357-a2c9-ee78e0e7' does not exist"
                },
                "status": {
                    "type": "string",
                    "example": "Resource not found"
                }
            }
        },
        "api.ErrorRemoteServerUnavailableExample": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Remote server error, please contact administrator"
                },
                "status": {
                    "type": "string",
                    "example": "Remote server unavailable"
                }
            }
        },
        "api.ErrorRenderExample": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "rendering error"
                },
                "status": {
                    "type": "string",
                    "example": "Error rendering response"
                }
            }
        },
...
        "dto.ProblemOfTarget": {
            "type": "object",
            "properties": {
                "argument": {
                    "type": "string",
                    "example": "consolidateBy"
                },
                "description": {
                    "type": "string",
                    "example": "This function affects only visual graph representation. It is meaningless in Moira"
                },
                "position": {
                    "type": "integer",
                    "example": 0
                },
                "problems": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ProblemOfTarget"
                    }
                },
                "type": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.typeOfProblem"
                        }
                    ],
                    "example": "warn"
                }
            }
        },
...
        "dto.SaveTriggerResponse": {
            "type": "object",
            "properties": {
                "checkResult": {
                    "$ref": "#/definitions/dto.TriggerCheckResponse"
                },
                "id": {
                    "type": "string",
                    "example": "trigger_id"
                },
                "message": {
                    "type": "string",
                    "example": "trigger created"
                }
            }
        },
...
        "dto.TreeOfProblems": {
            "type": "object",
            "properties": {
                "syntax_ok": {
                    "type": "boolean",
                    "example": true
                },
                "tree_of_problems": {
                    "$ref": "#/definitions/dto.ProblemOfTarget"
                }
            }
        },
        "dto.Trigger": {
            "type": "object",
            "properties": {
                "alone_metrics": {
                    "description": "A list of targets that have only alone metrics",
                    "type": "object",
                    "additionalProperties": {
                        "type": "boolean"
                    },
                    "example": {
                        "t1": true
                    }
                },
                "created_at": {
                    "description": "Datetime when the trigger was created",
                    "type": "string"
                },
                "created_by": {
                    "description": "Username who created trigger",
                    "type": "string"
                },
                "desc": {
                    "description": "Description string",
                    "type": "string",
                    "example": "check the size of /var/log"
                },
                "error_value": {
                    "description": "ERROR threshold",
                    "type": "number",
                    "example": 1000
                },
                "expression": {
                    "description": "Used if you need more complex logic than provided by WARN/ERROR values",
                    "type": "string",
                    "example": ""
                },
                "id": {
                    "description": "Trigger unique ID",
                    "type": "string",
                    "example": "292516ed-4924-4154-a62c-ebe312431fce"
                },
                "is_remote": {
                    "description": "Shows if trigger is remote (graphite-backend) based or stored inside Moira-Redis DB",
                    "type": "boolean",
                    "example": false
                },
                "mute_new_metrics": {
                    "description": "If true, first event NODATA → OK will be omitted",
                    "type": "boolean",
                    "example": false
                },
                "name": {
                    "description": "Trigger name",
                    "type": "string",
                    "example": "Not enough disk space left"
                },
                "patterns": {
                    "description": "Graphite patterns for trigger",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        ""
                    ]
                },
                "sched": {
                    "description": "Determines when Moira should monitor trigger",
                    "allOf": [
                        {
                            "$ref": "#/definitions/moira.ScheduleData"
                        }
                    ]
                },
                "tags": {
                    "description": "Set of tags to manipulate subscriptions",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "server",
                        "disk"
                    ]
                },
                "targets": {
                    "description": "Graphite-like targets: t1, t2, ...",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "devOps.my_server.hdd.freespace_mbytes"
                    ]
                },
                "throttling": {
                    "type": "integer",
                    "example": 0
                },
                "trigger_type": {
                    "description": "Could be: rising, falling, expression",
                    "type": "string",
                    "example": "rising"
                },
                "ttl": {
                    "description": "When there are no metrics for trigger, Moira will switch metric to TTLState state after TTL seconds",
                    "type": "integer",
                    "example": 600
                },
                "ttl_state": {
                    "description": "When there are no metrics for trigger, Moira will switch metric to TTLState state after TTL seconds",
                    "type": "string",
                    "example": "NODATA"
                },
                "updated_at": {
                    "description": "Datetime  when the trigger was updated",
                    "type": "string"
                },
                "updated_by": {
                    "description": "Username who updated trigger",
                    "type": "string"
                },
                "warn_value": {
                    "description": "WARN threshold",
                    "type": "number",
                    "example": 500
                }
            }
        },
...
        "dto.TriggerCheckResponse": {
            "type": "object",
            "properties": {
                "targets": {
                    "description": "Graphite-like targets: t1, t2, ...",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.TreeOfProblems"
                    }
                }
            }
        },
...
        "dto.TriggersList": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/moira.TriggerCheck"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "pager": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "moira.ScheduleData": {
            "type": "object",
            "properties": {
                "days": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/moira.ScheduleDataDay"
                    }
                },
                "endOffset": {
                    "type": "integer",
                    "example": 1439
                },
                "startOffset": {
                    "type": "integer",
                    "example": 0
                },
                "tzOffset": {
                    "type": "integer",
                    "example": -60
                }
            }
        },
        "moira.ScheduleDataDay": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "type": "string",
                    "example": "Mon"
                }
            }
        },
...
    },
...
}